Убрал лишние файлы, подготовил к основному ветке main
