+++ lib/main.dart
@@ -1,14 +1,11 @@
-import 'package:chirk/activities/test.dart';
+import 'package:chirk/pages/homePage.dart';
 import 'package:chirk/theme/theme.dart';
 import 'package:flutter/material.dart';
 
-import 'activities/home.dart';
-
 void main() => runApp(MaterialApp(
-  theme: createLightYellowTheme(),
-  initialRoute: '/',
-  routes: {
-    //'/': (context) => Splash(),
-    '/': (context) => TestActivity(),
-  },
-));
+      theme: lightYellowTheme(),
+      initialRoute: '/',
+      routes: {
+        '/': (context) => HomePage(),
+      },
+    ));
