+++ lib/templates/abstractBody.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+abstract class AbstractBody {
+  String title;
+
+  AbstractBody(this.title);
+
+  Widget createBody();
+}
