+++ lib/templates/abstractStandartState.dart
@@ -1,18 +1,21 @@
-import 'package:chirk/activities/new%20test/interfaceTemplate.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-abstract class AbstractStandardState extends State<StatefulWidget> implements InterfaceTemplate{
+import 'abstractBody.dart';
 
+abstract class AbstractNavigationState extends State<StatefulWidget>
+    implements AbstractBody {
   @override
   String title;
-  AbstractStandardState(this.title);
+
+  AbstractNavigationState(this.title);
 
   @override
   Widget build(BuildContext context) {
-    return standardNavigation(context, title);
+    return navigationTemplate(context, title);
   }
-  Widget standardNavigation(BuildContext context, String title) {
+
+  Widget navigationTemplate(BuildContext context, String title) {
     return Scaffold(
       appBar: AppBar(
         title: Title(
@@ -23,7 +26,8 @@ abstract class AbstractStandardState extends State<StatefulWidget> implements In
       body: createBody(),
       bottomNavigationBar: BottomNavigationBar(
         items: const <BottomNavigationBarItem>[
-          BottomNavigationBarItem(icon: Icon(Icons.list), label: "Лента чирков"),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.list), label: "Лента чирков"),
           BottomNavigationBarItem(icon: Icon(Icons.add), label: "Создать чирк"),
           BottomNavigationBarItem(icon: Icon(Icons.person), label: "Профиль"),
         ],
