+++ lib/theme/theme.dart
@@ -1,17 +1,14 @@
 import 'package:flutter/material.dart';
 
-ThemeData createLightYellowTheme(){
+ThemeData lightYellowTheme() {
   return ThemeData(
-    brightness: Brightness.light,
-    primarySwatch: Colors.yellow,
-    scaffoldBackgroundColor: Colors.yellow[50],
-    bottomAppBarTheme: const BottomAppBarTheme(
-      color: Colors.black,
-    ),
-    canvasColor: Colors.yellow[50],
-  iconTheme: IconThemeData(
-
-  )
+      brightness: Brightness.light,
+      primarySwatch: Colors.yellow,
+      scaffoldBackgroundColor: Colors.yellow[50],
+      bottomAppBarTheme: const BottomAppBarTheme(
+        color: Colors.black,
+      ),
+      canvasColor: Colors.yellow[50],
+      iconTheme: IconThemeData()
   );
-
 }
